+++ packages/core/src/graph/graph.ts
@@ -1240,7 +1240,7 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
 
     const { nodes = [], edges = [], combos = [] } = data;
 
-    this.clear();
+    this.clear(true);
 
     this.emit('beforerender');
 
@@ -1277,7 +1277,7 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
       if (self.get('fitCenter')) {
         self.fitCenter();
       }
-      self.emit('afterrender');
+      self.emit('afterrender', {message: 'layout controller do not exist'});
       self.set('animate', animate);
     }
     // 将在 onLayoutEnd 中被调用
@@ -1289,7 +1289,7 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
         self.fitCenter();
       }
       self.autoPaint();
-      self.emit('afterrender');
+      self.emit('afterrender', {message: 'layout success'});
       if (self.get('fitView') || self.get('fitCenter')) {
         self.set('animate', animate);
       }
@@ -1341,7 +1341,7 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
     const self = this;
     let item: INode;
     const itemMap: NodeMap = this.get('itemMap');
-
+    
     each(models, (model) => {
       item = itemMap[model.id];
       if (item) {
@@ -2139,7 +2139,7 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
    * 清除画布元素
    * @return {object} this
    */
-  public clear(): AbstractGraph {
+  public clear(avoidEmit: boolean = false): AbstractGraph {
     const canvas: ICanvas = this.get('canvas');
     canvas.clear();
 
@@ -2147,7 +2147,7 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
 
     // 清空画布时同时清除数据
     this.set({ itemMap: {}, nodes: [], edges: [], groups: [], combos: [], comboTrees: [] });
-    this.emit('afterrender');
+    if (!avoidEmit) this.emit('afterrender');
     return this;
   }
 
@@ -2872,7 +2872,7 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
       hullMap = {};
       this.set('hullMap', hullMap);
     }
-    if (!parent) {
+    if (!parent || parent.get('destroyed')) {
       parent = this.get('group').addGroup({
         id: 'hullGroup',
       });
@@ -2897,9 +2897,9 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
 
   /**
    * 获取当前 graph 中存在的包裹轮廓
-   * @return {[key: string]: Hull} hullId 对应的 hull 实例
+   * @return {[key: string]: Hull} Hull 的 map，hullId 对应的 hull 实例
    */
-  public getHulls(): Hull[] {
+  public getHulls(): {[key: string]: Hull} {
     return this.get('hullMap');
   }
 
@@ -2926,9 +2926,10 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
   public removeHulls() {
     const hulls = this.getHulls();
     if (!hulls || !Object.keys(hulls).length) return;
-    hulls.forEach((hull) => {
+    Object.keys(hulls).forEach(key => {
+      const hull = hulls[key];
       hull.destroy();
-    });
+    })
     this.set('hullMap', {});
   }
 }